Added stubbed unit tests and improvements to the logging system
Due to the fact that i've been unable to fine an acceptable solution for
adding mocks to the GameState unit tests i've decided to stub out some
tests for now until I wish to revisit the issue. The current problem
is due to the fact that mocking Ogre would require a large amount of
effort and running it normally would be having to create a render
window which is something I don't want to do for the unit tests.

I've made some additional improvements to GameRoot by finding the users
home folder, creating a hardwar directory in there that can then be used
for user configurations and logging. Ogre no longer spams the console
with the log anymore.